Replace per-transaction unit conversion with pack relationships
Found live, manually testing the unit selector shipped last commit: even
with instructions it was confusing, and it couldn't represent a real
case -- a box of meat sold either whole or in any loose kg amount.
Redesigned rather than patched.

Two independent products (a case, a can) can now be linked by a declared
PackRelationship ("1 Coke Case = 24 Coke Can"), converted between only
by an explicit "repack" action (app/inventory.py's record_repack) that
writes two linked stock_movements atomically -- never inferred at sale
time, since whether to crack open a sealed case is the owner's judgment
call. The meat/box-of-10.5kg scenario needs nothing new beyond this: two
linked products, same mechanism.

This surfaced a real requirement: quantities need to be Decimal, not
Integer, since a fractional sale (2.3 kg) has to fit the same shared
column every counted product's quantity uses. Widened
sales.quantity/stock_movements.quantity_delta/reorder_level everywhere
(migration af491567128d), including WhatsApp/chat's tool schemas, not
just the web entry form.

Caught along the way, by the full test suite rather than inspection:
report_metrics.py and forecasting.py were both passing a raw Decimal
straight into dicts that app/agents.py hands to json.dumps() for LLM
prompts -- every report/insights run would have started crashing the
moment any product had a low-stock item. Fixed by routing through the
_to_float() helper both modules already use for every other number.

Removed rather than left as dead code: ProductUnit's endpoints, the unit
dropdown on /entry and /inventory's Adjust form, to_base_quantity. The
product_units table itself stays, unused, per the schema-frozen rule.

See docs/decisions.md's [2026-09-14] entry for the full reasoning.

+++ apps/api/app/models/records.py
@@ -15,7 +15,10 @@ class Sale(Base):
     sale_date = Column(Date, nullable=True)
     product_name = Column(String, nullable=True)
     category = Column(String, nullable=True)
-    quantity = Column(Integer, nullable=True)
+    # Numeric, not Integer: one shared column across every sale, and some
+    # products (loose meat/produce by weight) sell in fractional amounts.
+    # See docs/decisions.md [2026-09-14].
+    quantity = Column(Numeric, nullable=True)
     unit_price = Column(Numeric, nullable=True)
     discount = Column(Numeric, nullable=True)
     total_amount = Column(Numeric, nullable=True)
@@ -40,8 +43,11 @@ class Inventory(Base):
     upload_session_id = Column(UUID(as_uuid=True), ForeignKey("upload_sessions.id"), nullable=False, index=True)
     product_name = Column(String, nullable=True)
     category = Column(String, nullable=True)
-    quantity = Column(Integer, nullable=True)
-    reorder_level = Column(Integer, nullable=True)
+    # Widened alongside Sale.quantity/Product.reorder_level -- an inventory
+    # row for a weight-tracked product (a 10.5 kg box) would otherwise
+    # silently truncate on insert. See docs/decisions.md [2026-09-14].
+    quantity = Column(Numeric, nullable=True)
+    reorder_level = Column(Numeric, nullable=True)
     supplier = Column(String, nullable=True)
     supplier_id = Column(UUID(as_uuid=True), ForeignKey("suppliers.id"), nullable=True, index=True)
     cost_price = Column(Numeric, nullable=True)
